refactor(wordle): Move validation logic to src/backend/
